decimali

+++ Classes/BasicStrategyS17.cs
@@ -8,7 +8,7 @@ namespace Classes
     [Serializable]
     public class BasicStrategyS17 : StrategiaGiocatore
     {
-        public override Giocatore.Puntata Strategy(Giocatore giocatore, Mazziere mazziere, decimal conteggio)
+        public override Giocatore.Puntata Strategy(Giocatore giocatore, Mazziere mazziere, double conteggio)
         {
             int pg= giocatore.Punteggio;
             int pm = mazziere.Carte.First().Valore;
@@ -109,7 +109,7 @@ namespace Classes
             }
         }
 
-        public override double Puntata(Giocatore giocatore, double puntataMinima, double puntataBase, int Conteggio)
+        public override double Puntata(Giocatore giocatore, double puntataMinima, double puntataBase, double Conteggio)
         {
             if (Conteggio <= -2)
                 return puntataMinima;
@@ -133,7 +133,7 @@ namespace Classes
             return Conteggio;
         }
 
-        public override bool Assicurazione(Giocatore giocatore, decimal conteggio)
+        public override bool Assicurazione(Giocatore giocatore, double conteggio)
         {
             return false;
         }
